test(git): align tests with source files
